Added path variable for number of steps, updated JSON
When using the /TempSim/Step path to run the simulation, you should now include the number of iterations to run; ie. /TempSim/Step/5 to run 5 iterations. The state of all agents will be shown in the browser and maintained throughout subsequent calls to the route. To specifically view the locations and actions in the system, use the routes /Location/Print and /Action/Print, respectively.