Initial implementation of HistoryLogger with MongoDB and very surface-level testing
The HistoryLogger is abstract and generic like Reality and Knowledge Sims for alternative implementations, but the default/local implementation is set up for a local MongoDB server with a database titled "SimManager".

Current functionality includes logging timesteps and NPC states when their action or a motive changes (generally these will occur at the same time, when an action ends and a motivation is updated), and saving full simulation states to a collection of choice.

Planned functionality includes loading the SimManager from a saved state. This should afford the branching timeline capability depending on how the frontend chooses to utilize it